+++ src/gleis/gleise/update.rs
@@ -14,12 +14,12 @@ use log::error;
 use parking_lot::Mutex;
 
 use crate::{
+    application::steuerung::{AsyncAktualisieren, MitSteuerung, Steuerung},
     gleis::{
         gerade::Gerade,
         gleise::{
             daten::{Gleis, GleiseDaten, RStern},
             id::{mit_any_id, AnyId, AnyIdRef, GleisIdRef, StreckenabschnittIdRef},
-            steuerung::{AsyncAktualisieren, MitSteuerung, Steuerung},
             Gehalten, Gleise, ModusDaten, Nachricht,
         },
         kreuzung::Kreuzung,
